test_requirements.txt
-r requirements.txt
pytest
pytest-django
freezegun
